// src/config/enums/configurationTypes.enum.ts
export enum ConfigurationTypes {
  PORT = 'PORT',

  POSTGRES_HOST = 'POSTGRES_HOST',
  POSTGRES_PORT = 'POSTGRES_PORT',
  POSTGRES_USER = 'POSTGRES_USER',
  POSTGRES_PASSWORD = 'POSTGRES_PASSWORD',
  POSTGRES_DB = 'POSTGRES_DB',

  BCRYPT_SALT_OR_ROUNDS = 'BCRYPT_SALT_OR_ROUNDS',

  JWT_ACCESS_SECRET = 'JWT_ACCESS_SECRET',
  JWT_REFRESH_SECRET = 'JWT_REFRESH_SECRET',
  JWT_ACCESS_EXPIRATION_TIME = 'JWT_ACCESS_EXPIRATION_TIME',
  JWT_REFRESH_EXPIRATION_TIME = 'JWT_REFRESH_EXPIRATION_TIME',

  SMTP_HOST = 'SMTP_HOST',
  SMTP_PORT = 'SMTP_PORT',
  SMTP_USER = 'SMTP_USER',
  SMTP_PASSWORD = 'SMTP_PASSWORD',
}